# improve lexing unquoted and number literals