```console
$ benchstat a.txt b.txt      
goos: darwin
goarch: arm64
pkg: go.expect.digital/mf2/parse
cpu: Apple M2 Max
                         │    a.txt     │                b.txt                │
                         │    sec/op    │   sec/op     vs base                │
ComplexMessage_String-12    879.8n ± 1%   894.2n ± 1%   +1.64% (p=0.000 n=10)
Lex-12                     1308.0n ± 1%   974.2n ± 1%  -25.52% (p=0.000 n=10)
Parse-12                    3.058µ ± 0%   2.700µ ± 0%  -11.69% (p=0.000 n=10)
geomean                     1.521µ        1.330µ       -12.56%

                         │    a.txt     │                 b.txt                  │
                         │     B/op     │     B/op      vs base                  │
ComplexMessage_String-12     776.0 ± 0%     776.0 ± 0%        ~ (p=1.000 n=10) ¹
Lex-12                       656.0 ± 0%     112.0 ± 0%  -82.93% (p=0.000 n=10)
Parse-12                   7.242Ki ± 0%   6.711Ki ± 0%   -7.34% (p=0.000 n=10)
geomean                    1.521Ki          842.1       -45.92%
¹ all samples are equal

                         │    a.txt    │                b.txt                 │
                         │  allocs/op  │ allocs/op   vs base                  │
ComplexMessage_String-12    33.00 ± 0%   33.00 ± 0%        ~ (p=1.000 n=10) ¹
Lex-12                     19.000 ± 0%   3.000 ± 0%  -84.21% (p=0.000 n=10)
Parse-12                    53.00 ± 0%   37.00 ± 0%  -30.19% (p=0.000 n=10)
geomean                     32.15        15.42       -52.05%
```